+++ components/Employee/Activator/Base.vue
@@ -2,7 +2,7 @@
 /*****************************************************************************
  * @file ./components/Employee/Activator/Base.vue
  * @description 従業員の基本情報表示コンポーネント
- * - `EmployeeEditor` の activator スロット用コンポーネント
+ * - `EmployeeManager` の activator スロット用コンポーネント
  *****************************************************************************/
 import dayjs from "dayjs";
 import { Employee } from "@/schemas";